Use test from expired data to replace tests from delivered data
